requirements.txt
aiohttp>=3.8.5
aiosignal>=1.3.1
APScheduler>=3.10.1
asgiref>=3.7.2
astroid>=2.15.6
async-timeout>=4.0.3
attrs>=23.1.0
autopep8>=2.0.2
bcrypt>=4.0.1
bleach>=6.0.0
blinker>=1.6.2
certifi>=2023.7.22
cffi>=1.15.1
chardet>=5.2.0
charset-normalizer>=3.2.0
click>=8.1.6
colorama>=0.4.6
cryptography>=41.0.3
cssselect>=1.2.0
decorator>=5.1.1
dicttoxml>=1.7.16
dill>=0.3.7
Django>=4.2.4
django-extensions>=3.2.3
django-google-maps>=0.13.0
django-rest-framework>=0.1.0
djangorestframework>=3.14.0
docutils>=0.20.1
drf-generators>=0.5.0
drf-nested-routers>=0.93.4
exceptiongroup>=1.1.2
Flask>=2.3.2
frozenlist>=1.4.0
future>=0.18.3
geocoder>=1.38.1
geoip2>=4.7.0
gevent>=23.7.0
googlemaps>=4.10.0
greenlet>=2.0.2
h11>=0.14.0
idna>=3.4
importlib-metadata>=6.8.0
ip2geotools>=0.1.6
IP2Location>=8.10.0
isort>=5.12.0
itsdangerous>=2.1.2
jaraco.classes>=3.3.0
Jinja2>=3.1.2
keyring>=24.2.0
lazy-object-proxy>=1.9.0
lxml>=4.9.3
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
maxminddb>=2.4.0
mccabe>=0.7.0
mdurl>=0.1.2
meraki>=1.36.0
more-itertools>=10.1.0
multidict>=6.0.4
outcome>=1.2.0
packaging>=23.1
paramiko>=3.3.1
pip-review>=1.3.0
pkginfo>=1.9.6
platformdirs>=3.10.0
pproxy>=2.7.8
pycodestyle>=2.11.0
pycparser>=2.21
Pygments>=2.16.1
pylint>=2.17.5
PyNaCl>=1.5.0
pyparsing>=3.1.1
pyquery>=2.0.0
PySocks>=1.7.1
pytz>=2023.3
ratelim>=0.1.6
readme-renderer>=40.0
requests>=2.31.0
requests-toolbelt>=1.0.0
rfc3986>=2.0.0
rich>=13.5.2
selenium>=4.11.2
six>=1.16.0
sniffio>=1.3.0
sortedcontainers>=2.4.0
sqlparse>=0.4.4
toml>=0.10.2
tomli>=2.0.1
tomlkit>=0.12.1
tqdm>=4.66.1
trio>=0.22.2
trio-websocket>=0.10.3
twine>=4.0.2
typed-ast>=1.5.5
typing>=3.7.4.3
typing_extensions>=4.7.1
tzlocal>=5.0.1
urllib3>=2.0.4
webencodings>=0.5.1
websocket-client>=1.6.1
websocket-server>=0.6.4
Werkzeug>=2.3.6
wrapt>=1.15.0
wsproto>=1.2.0
yarl>=1.9.2
zipp>=3.16.2
zope.event>=5.0
zope.interface>=6.0
flask>=3.0.2
